Refactored game state into a context and hook
