+++ src/context/theme.ts
@@ -1,13 +1,29 @@
-import { createContext } from 'react'
-import { dark } from '../styles/colors'
-import { IthemeContext } from '../types/theme'
-import { Itheme } from '../types/styles'
+import { useCallback, useState } from "react"
+import { useLocalStorage } from "react-use"
+import { dark, light } from "../styles/colors"
+import { Itheme } from '../types/styles';
+import { buildGenericContext } from './genericContext';
 
-export const default_theme_context = {
-    isDark: true,
-    toggleTheme: () => console.warn('No toggle theme function defined.'),
+
+const useTheme = () => {
+    const [isDark, setIsDark] = useLocalStorage<boolean>('isDark', true)
+    const [theme, setTheme] = useState<Itheme>(dark)
+
+    const toggleTheme = useCallback(() => {
+        console.log(`Toggle theme`)
+
+        let newTheme: Itheme
+        if (!isDark)
+            newTheme = dark
+        else
+            newTheme = light
+
+        setTheme(newTheme)
+        setIsDark(!isDark)
+
+    }, [setIsDark, isDark, setTheme])
+
+    return { theme, isDark: !!isDark, toggleTheme }
 }
 
-export const CurrentTheme = createContext<Itheme>(dark)
-
-export const ThemeContext = createContext<IthemeContext>(default_theme_context)
+export const [ThemeProvider, useThemeContext] = buildGenericContext(useTheme)
