--- src/hooks/useTheme.ts
@@ -1,29 +1,0 @@
-import { useCallback, useState } from "react"
-import { useLocalStorage } from "react-use"
-import { dark, light } from "../styles/colors"
-import { Itheme } from '../types/styles';
-import { buildGenericContext } from '../context/genericContext';
-
-
-const useTheme = () => {
-    const [isDark, setIsDark] = useLocalStorage<boolean>('isDark', true)
-    const [theme, setTheme] = useState<Itheme>(dark)
-
-    const toggleTheme = useCallback(() => {
-        console.log(`Toggle theme`)
-
-        let newTheme: Itheme
-        if (!isDark)
-            newTheme = dark
-        else
-            newTheme = light
-
-        setTheme(newTheme)
-        setIsDark(!isDark)
-
-    }, [setIsDark, isDark, setTheme])
-
-    return { theme, isDark: !!isDark, toggleTheme }
-}
-
-export const [ThemeProvider, useThemeContext] = buildGenericContext(useTheme)
